move onWalletComponentsUpdated to useLauncher
